Consolidate TestContext builder to always produce a PendingAsyncTestContext.
This will reduce the complexity if we need to apply any further
transformations after the target is resolved. Since we finish
resolving the target in one place, we only have to worry about one
location when applying the transformation.

If the target is immediately resolved, the listener on the future will
still run and should immediately replace the pending context with the
actual target.

This should be a no-op change.

PiperOrigin-RevId: 250532653